Add safe-guard to prepare_dev_server_config.py

+++ tools/prepare_dev_server_config.py
@@ -11,7 +11,12 @@ _FOLDER = Path(__file__).parent
 
 
 def main():
-    with _FOLDER.joinpath("dev-server-configuration.json").open("w") as f:
+    target = _FOLDER.joinpath("dev-server-configuration.json")
+    if target.is_file():
+        print(f"{target} already exists, not updated")
+        return
+
+    with target.open("w") as f:
         json.dump(
             {
                 "server_address": "http://127.0.0.1:5000",
